@@ -344,7 +344,7 @@ fn do_status(filename: &str, args: &[String]) {
                     pubref,
                     key.privref().unwrap_or_default(),
                 );
-                println!("\t\tState: {}", keystate);
+                println!("\t\tState: {keystate}");
             }
             KeyType::Csk(keystate_ksk, keystate_zsk) => {
                 println!(
@@ -352,8 +352,8 @@ fn do_status(filename: &str, args: &[String]) {
                     pubref,
                     key.privref().unwrap_or_default(),
                 );
-                println!("\t\tKSK role state: {}", keystate_ksk,);
-                println!("\t\tZSK role state: {}", keystate_zsk,);
+                println!("\t\tKSK role state: {keystate_ksk}");
+                println!("\t\tZSK role state: {keystate_zsk}");
             }
         }
         let ts = key.timestamps();
@@ -447,7 +447,7 @@ fn report_actions(actions: Result<Vec<Action>, Error>, ks: &KeySet) {
                         | KeyType::Include(keystate) => keystate,
                     };
                     if status.present() {
-                        println!("\t\t\t{}", pubref);
+                        println!("\t\t\t{pubref}");
                     }
                 }
                 println!("\t\tKeys signing the DNSKEY RRset:");
@@ -456,7 +456,7 @@ fn report_actions(actions: Result<Vec<Action>, Error>, ks: &KeySet) {
                         KeyType::Ksk(keystate)
                         | KeyType::Csk(keystate, _) => {
                             if keystate.signer() {
-                                println!("\t\t\t{}", pubref);
+                                println!("\t\t\t{pubref}");
                             }
                         }
                         KeyType::Zsk(_) | KeyType::Include(_) => (),
@@ -471,7 +471,7 @@ fn report_actions(actions: Result<Vec<Action>, Error>, ks: &KeySet) {
                         KeyType::Zsk(keystate)
                         | KeyType::Csk(_, keystate) => {
                             if keystate.signer() {
-                                println!("\t\t{}", pubref);
+                                println!("\t\t{pubref}");
                             }
                         }
                         KeyType::Ksk(_) | KeyType::Include(_) => (),
@@ -489,7 +489,7 @@ fn report_actions(actions: Result<Vec<Action>, Error>, ks: &KeySet) {
                         | KeyType::Include(keystate) => keystate,
                     };
                     if status.at_parent() {
-                        println!("\t\t{}", pubref);
+                        println!("\t\t{pubref}");
                     }
                 }
             }
@@ -504,7 +504,7 @@ fn report_actions(actions: Result<Vec<Action>, Error>, ks: &KeySet) {
                         | KeyType::Include(keystate) => keystate,
                     };
                     if status.at_parent() {
-                        println!("\t\t{}", pubref);
+                        println!("\t\t{pubref}");
                     }
                 }
             }
